[Improvement][api] Introduce resource, scheduler, taskinstance interface
